T5096: firewall: Change 'accept' rule action from 'return' to 'accept'
I was expecting a `jump` action was designed to use with a chain that has `return` default
action to create reusable rule sets.

e.g.

```
interface eth8 {
     in {
         name stdnt-in
     }
 }
 name floating {
     default-action return
     enable-default-log
     rule 1 {
         action accept
         destination {
             address 10.236.253.11
         }
         log enable
         protocol icmp
     }
 }
 name stdnt-in {
     default-action drop
     enable-default-log
     rule 1 {
         action jump
         jump-target floating
         log enable
     }
 }
```

With the above configuration, it should accept ingress ICMP packets to
destination 10.236.253.11. However this doesn't work. Those ICMP packets
will be dropped because Vyos converts `accept` action to `return` action
for nftables. The rendered nftables rules are look like:

```
        chain NAME_floating {
                meta l4proto icmp ip daddr 10.236.253.11 log prefix "[floating-1-A]" counter packets 0 bytes 0 return comment "floating-1"
                counter packets 0 bytes 0 log prefix "[floating-default-R]" return comment "floating default-action return"
        }

        chain NAME_stdnt-in {
                log prefix "[stdnt-in-1-J]" counter packets 0 bytes 0 jump NAME_floating comment "stdnt-in-1"
                counter packets 0 bytes 0 log prefix "[stdnt-in-default-D]" drop comment "stdnt-in default-action drop"
        }

        chain VYOS_FW_FORWARD {
        	type filter hook forward priority 0; policy accept;
 		iifname eth8 counter jump NAME_stdnt
                # here we have rules for other interfaces
                jump VYOS_POST_FW
        }

```

Maybe I have missed something, but it doesn't seem to have a good reason
to replace the `accept` action with `return` action. This replacement
actually makes the `jump` action useless because you can't actually
accept a rule in a sub-chain.

Also because of the `return` action, the `VYOS_POST_FW` chain is always
be run. However, that chain is hardcoded to be empty so it does nothing.

So I would like to change it. In order there is a good reason to do
further processing after a rule is accepted, we have have other
workarounds.